requirements.txt
feedparser==5.2.1
python-dateutil==2.6.0
six==1.10.0
